Move opencodelists.py to main jobserver app
